+++ src/presenter/films-presenter.js
@@ -6,8 +6,7 @@ import FilmListTitleView from '../view/film-list-title-view';
 import FilmListContainerView from '../view/film-list-container-view';
 import FilmItemView from '../view/film-item-view';
 import ShowMoreButtonView from '../view/show-more-button-view';
-import FilmTopRateView from '../view/film-top-rate-view';
-import FilmMostCommentView from '../view/film-most-comment-view';
+import FilmMostView from '../view/film-most-view';
 
 const FILMS_NUMBER = 5;
 
@@ -24,8 +23,8 @@ export default class FilmsPresenter {
     render(new FilmListTitleView(), this.filmList.getElement());
     render(this.filmListContainer, this.filmList.getElement());
     render(new ShowMoreButtonView(), this.filmList.getElement());
-    render(new FilmTopRateView(), this.filmSection.getElement());
-    render(new FilmMostCommentView(), this.filmSection.getElement());
+    render(new FilmMostView('MostRated'), this.filmSection.getElement());
+    render(new FilmMostView('MostCommented'), this.filmSection.getElement());
 
     for (let i = 0; i < FILMS_NUMBER; i++) {
       render(new FilmItemView(), this.filmListContainer.getElement());
