+++ src/view/film-list-view.js
@@ -1,6 +1,10 @@
 import {createElement} from '../render';
 
-const createFilmListTemplate = () => '<section class="films-list"></section>';
+const createFilmListTemplate = () => (
+  `<section class="films-list">
+    <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+  </section>`
+);
 
 export default class FilmListView {
   getTemplate() {
